requirements.txt
discord.py==2.3.2
PyYAML>=6.0 # For reading config.yaml
aiohttp>=3.9.0 # For async HTTP requests in llm_api
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dateutil>=2.8.0 # For flexible date/time parsing